refatorando o código e criando função para formatar datas
